% Generated by roxygen2 (4.0.0): do not edit by hand
\name{CheckLocusNames}
\alias{CheckLocusNames}
\title{Check for locus names that are not consistent between the two alleles of the same locus}
\usage{
CheckLocusNames(D)
}
\arguments{
  \item{D}{A data frame in which the loci start in column 1
  and the column for the second allele at a locus named
  "LocusX" is "LocusX.1".  Thus, it is the sort of data
  frame that would result from reading in a two-column
  (ToolKit) format file while setting the first column to
  be the row names, and using R's default method for making
  column names unique}
}
\value{
A matrix of the two instances of each locus name that are not identical in the
file.
}
\description{
Check for locus names that are not consistent between the two alleles of the same locus
}
\details{
Note that if someone already added ".1" to the second occurrence of each locus name, then
that will work fine as R will already find it to be unique (however adding ".a" or ".b" would break things).
}
\examples{
silly.data <- data.frame(LocA=c(2,2), LocA=c(3,2), LocB=c(45,42), LocBB=c(48,42))
  rownames(silly.data) <- c("Fish1", "Fish2")
  silly.data  # just look at it. Notice that LocB and LocBB aren't right!
  CheckLocusNames(silly.data)

  # note that the example steelhead data are OK
  CheckLocusNames(sthd.geno.A)
}

