% Generated by roxygen2 (4.0.0): do not edit by hand
\name{ExcelColumns}
\alias{ExcelColumns}
\title{give the column header (A or J or AC, etc) name in an excel file for}
\usage{
ExcelColumns(C, Shift = 0)
}
\arguments{
  \item{C}{a vector of integers giving the column numbers}

  \item{Shift}{the amount to add to C before returning the
  column name. This is useful when columns were stripped
  from the Excel file to get it into an R data frame.  For
  example, if there were 15 columns in the Excel file that
  were dropped before reading the remaining columns, Shift
  would be 15}
}
\description{
give the column header (A or J or AC, etc) name in an excel file for
}
\details{
As currently implemented, this can give values out to 18278 columns
}
\examples{
ExcelColumns(11:18)
  ExcelColumns(11:18, Shift=10)
}

