% Generated by roxygen2 (4.0.0): do not edit by hand
\name{FindLocusColumns}
\alias{FindLocusColumns}
\title{Find the excel columns holding specific loci.}
\usage{
FindLocusColumns(X, N, Shift = 1)
}
\arguments{
  \item{X}{names of loci you want to try to find in the
  file}

  \item{N}{names of the loci in the order they appear in in
  the excel file}

  \item{Shift}{the number of the column in the excel file
  that the first locus in N appears}
}
\description{
Given a vector X of locus names for which you want positions, and a vector of locus names N that are in
 an excel file somewhere this returns the column names a la letters
}

