% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{Omykiss.Standard.Loci}
\alias{Omykiss.Standard.Loci}
\title{Names of 115 steelhead loci with standardized names and order}
\format{A character vector of length 115}
\usage{
data(Omykiss.Standard.Loci)
}
\description{
These data are simply the names of the loci in the lab that we
are the current standard ones that we will be maintaining in our data
}
\keyword{datasets}

