% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{sthd.geno.A}
\alias{sthd.geno.A}
\title{Genotype data from steelhead version A}
\format{A data frame with 10 rows and 228 variables (which is 114 loci)}
\usage{
data(sthd.geno.A)
}
\description{
These data are simply a data frame of genetic data
}
\keyword{datasets}

