% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{sthd.geno.B}
\alias{sthd.geno.B}
\title{Genotype data from steelhead version B}
\format{A data frame with 9 rows and 240 variables (which is 120 loci)}
\usage{
data(sthd.geno.B)
}
\description{
These data are simply a data frame of genetic data
}
\keyword{datasets}

