% Generated by roxygen2 (4.0.0): do not edit by hand
\name{stringdist.handle.NAs}
\alias{stringdist.handle.NAs}
\title{do stringdist but handle NAs in a way that is appropriate to my task}
\usage{
stringdist.handle.NAs(x, y, ...)
}
\arguments{
  \item{x}{character vector}

  \item{y}{character vector of same length as x.}

  \item{...}{other arguments to pass to the function
  \code{stringdist}}
}
\value{
returns the string distance between the i-th element in x and the i-th element in y
  and returns a vector with length=length(x) containing NAs wherever x or y was NA.
}
\description{
It seems that stringdist returns NA when any strings being compared to
are NA.  I needed somethign else.
}
\examples{
stringdist.handle.NAs(c("Big", "Bad", "Bob", NA), c(NA, "Bed", "Bob", NA))
}

